refactor(writer_app): Split git_service.py into modular structure
- Create git_service/ package directory
- Extract base.py for initialization and utilities
- Extract commit.py for commit operations
- Extract history.py for history and diff operations
- Extract branch.py for branch operations
- Extract status.py for status operations
- Replace 394-line file with mixin-based architecture
